Implement config reload
Server will watch its config and restart on change.
Care is taken not to restart when config file is obviously bad
(this won't catch all the errors though).
Case where file goes missing and reappears is also handled gracefully.
Restarting on signal is not implemented because it's not necessary:
sending a signal tells the server that the file has changed. In this
case server watches the file and does not need to be told about it.