docker_auth/github: store tokens in google cloud storage
Added a new implementation of TokenDB that uses Google Cloud Storage as
backend for storing of the tokens.

This makes container independent from the file system of the container
or from the target system and is good alternative to mounted clustered
file systems to container.

Docs examples are updated regarding this change as specifying token_db
configuration now becomes alternate between local file and gcs.