[build] Allow overriding CA_BUNDLE as environment var
In order to be able to use a different certificate than `/etc/ssl/certs/ca-certificates.crt`,
for example on Mac OS X, this patch allows to set the path to a certificate in an environment variable.

Example:

    $ curl -O https://curl.haxx.se/ca/cacert.pem

    $ CA_BUNDLE=cacert.pem make ca-certificates.crt

The certificate is copied as `ca-certificates.crt`, so that the Dockerfile works regardless of the source filename.