[build] Use Alpine in Dockerfile in order to install packages inside a running container
In order to be able to run tools like `curl`, for example to debug and verify SSL connections,
the base distribution has been changed to Alpine Linux in the Dockerfile.

The size differences between Alpine and busybox are not so significant,
and the ability to install tools like `curl` or `vim` in the container
is attractive for debugging.