[authn] Added the fetching of Github teams into token `labels`
This patch adds the ability to:

* Fetch user's team in configured organization (the `fetchTeams` function), with suppor for pagination of resuls
* Store them as list of strings in the `TokenDBValue` struct
* Return them as `labels` in the `Authenticate` function

The motivation is to use the list of user's teams as `labels` in the ACL configuration, eg.:

    - match: {labels: {"teams": "developers"}}
        actions: ["*"]
        comment: Developers have access to everything

The Github API used, https://developer.github.com/v3/orgs/teams/#list-user-teams, is currently marked as
experimental, but it's currently the only way how to fetch the teams of a user.

The teams are fetched _only_ when a Github organization is configured in YAML, with the assumption that
it will be used in context of a company Docker registry, which eg. allows public access for pulling,
and private access for pushing the images.

Implements: https://github.com/cesanta/docker_auth/issues/191