[authn] Added a Redis-based `TokenDB` implementation
This patch adds an implementation of TokenDB which stores user tokens in Redis (http://redis.io).

The primary motivation is to enable operation in a distributed environment, where multiple
instances of the `auth_server` are running, and the default LevelDB-based implementation
stores tokens in a file on disk.

This was inspired by the `GCSTokenDB` implementation, which solves a similar problem,
but is tied to Google Cloud Storage, which would introduce latency when fetching tokens eg. from AWS.