[authn] Changed the visual style for the `/github_auth` HTML page and added a new template for the result
This patch improves the visual style of the `/github_auth` page, adding some CSS and better HTML structure.
The Github logo is added as an inline SVG string in the CSS declaration.

It also adds a new HTML page for the subsequent page, when user has been authenticated at Github,
showing a formatted `docker login` command.

It also changes the handler in `server.go` to redirect to `/github_auth` immediately, when it was
configured in the YAML file, to skip and unnecessary step.