[build] Speed up build-release with a separate make target
`make build-release` currently takes a long time because of `govendor sync -v`.
This is a time consuming step, even if `vendor/` is already bind-mounted in the Docker builder image.

This commit creates two new make targets, `pull-vendor-deps` and `build-release-fast`.
The former needs to be executed once to pull the packages listed in vendor.json.
Then building new releases just needs`make build-release-fast`.

The old `build-release` target remains as is in functionality, but
depends on the `pull-vendor-deps` target to reduce duplication.