Add authorization based on Github teams (#219)
* [authn] Added the fetching of Github teams into token `labels`

This patch adds the ability to:

* Fetch user's team in configured organization (the `fetchTeams` function), with suppor for pagination of resuls
* Store them as list of strings in the `TokenDBValue` struct
* Return them as `labels` in the `Authenticate` function

The motivation is to use the list of user's teams as `labels` in the ACL configuration, eg.:

    - match: {labels: {"teams": "developers"}}
        actions: ["*"]
        comment: Developers have access to everything

The Github API used, https://developer.github.com/v3/orgs/teams/#list-user-teams, is currently marked as
experimental, but it's currently the only way how to fetch the teams of a user.

The teams are fetched _only_ when a Github organization is configured in YAML, with the assumption that
it will be used in context of a company Docker registry, which eg. allows public access for pulling,
and private access for pushing the images.

Implements: https://github.com/cesanta/docker_auth/issues/191

This patch improves the visual style of the `/github_auth` page, adding some CSS and better HTML structure.
The Github logo is added as an inline SVG string in the CSS declaration.

It also adds a new HTML page for the subsequent page, when user has been authenticated at Github,
showing a formatted `docker login` command.

It also changes the handler in `server.go` to redirect to `/github_auth` immediately, when it was
configured in the YAML file, to skip and unnecessary step.