[authn] Added the `Labels` property to the `TokenDBValue` struct (#217)
This patch extends the `TokenDBValue` struct in order to allow storing "labels" associated
with a user directly in the data structure, ie. without a need for secondary storage.

The primary motivation is related to the possibility of storing user's Github teams
as `labels`, and using them in the ACL configuration.