Send the scope class as part of the token
As it is right now, it is not possible to authenticate plugin requests
using `docker-auth`. The docker distribution registry expects[0] a scope
with `Type` repository and `docker-auth` is sending `repository(plugin)`
as that's what it receives in the request.

Instead of sending `repository(plugin)` we should be sending
`repository` and setting the scope class to `plugin`. The default class
is `image`, which is why this works fine right now.

Fixes #269