Fix LevelDB yaml unmarshall error and update documentation (#389)
In order to ensure the project correctly validates authentication
methods in the config file that can use a token db, this commit modifies
the logic to ensure we check if all of the token methods are nil before
outputting an error. Previously, if the local filesystem token db method
was not nil and the other two (redis and google cloud storage) were nil,
we would return an error and the config would be considered invalid.

Additionally, this commit documents the correct LevelDB settings for the
config file. Prior to the addition of Bcrypt hashing cost, just using
`token_db` was acceptable in the configuration along with a string
representing the path. Given the swap to a struct instead of a string,
we need to update documentation concerning this.